Updated set of files for project.
Updated set of files for project. Missing is the fractal selection.
Update means to feed data into the pipeline using the select input.
Proper timing of the sync pulses is off. Add shift register with stages
equal the number of stages in the pipeline + other clocked logic before
the output.